Move ProductOverview to own file, fixed ovular render issue on styles
