+++ client/public/style.css
@@ -25,10 +25,11 @@ body {
 }
 
 .style {
-  max-width: 50px;
-  max-height: auto;
-  border-radius: 100%;
+  object-fit: cover;
+  height: 50px;
+  width: 50px;
   padding: 10px;
+  border-radius: 100%;
 }
 
 .styleContainer {
